Add style option for 'orientation' to support landscape mode.
This diff adds a new styles option in `styles.drawing.orientation`, with
possible values:
- portrait (default)
- landscape

All draw methods are updated to correctly calculate the position of
elements depending on the specified orientation.

One quirk of this is that typically the 6th (low e) string is shown at
the bottom of landscape neck diagrams.  This means simply swapping the
x,y positions of the elements does not work (that would place the low e
string at the top of the diagram.).

To handle this, string index values are translated to an effective
string number using a new helper.  The effective string number is then
used to calculate relevant offsets.